Remove autofocus from first placeholder field
An accessibility audit in Oct 2023 reported that
focusing the email address field when sending an
email was confusing for screen reader users. This
is because it stops the browser announcing that a
new page has loaded.

We noticed that the audit didn't report the same
for subsequent placeholders so assume the
autofocus behaviour didn't disrupt the expected
behaviour in the same way, presumably because you
have already started entering placeholders so the
information that a new page has loaded is less
important than entering each one.

Based on all that, this limits the autofocus so it
only applies to all placeholders that follow the
first one.