Preview letter branding with `img`, not `iframe`
The iframe we use to show letter brandings only has value as a holder
for the image of the letter PDF so should just be that image.

Presenting the branding example as a document (in an iframe) is
confusing for screen reader users when it has no actual content other
than the image.

The auditors testing the pages for setting email and letter branding
raised issues for this in the audit on Oct 9th.