Only ask for recipients once in one-off journey
If someone uses a placeholder which is the same as the key we use for
the recipient[1], we ask for the value twice, but only store it once.

This is even worse if the user has a the same placeholder with different
space/casing in the same template[2]. In this case we would ask for the
value once for each different variant of spacing/case, but still only
store the first one entered.

This commit normalises all the placeholders after merging them with the
special recipient keys.

This means that every instance of that placeholder will be populated in
the template straight away, and the user won’t be asked for it again.

1. For example, `email address`, `phone number` or `address line 1`,
   `address line 2`, etc
2. For example `email address` and `EMAILADDRESS`