show organisations as a separate view
It can be confusing to users that organisations show up under "Live
services". This wasn't such an issue when it was just platform admins
but we have lots of org users now. Separate orgs out into their own
section with its own heading.

to do this, tweak how the "show_heading" field is calculated in the
service - lists. Now we count how many headings we'll need in python
to simplify the jinja.