Remove onevoicewales.wales from top level email domains file
Not sure why this one was added. Appears to see very little use
with only one user and doesn't appear to be either
- a top level domain (like gov.uk)
- a domain that is shared between organisations (like hscni.net)

Therefore I've made a new organisation for One Voice Wales and
given it this domain
https://www.notifications.service.gov.uk/organisations/bfd2c499-7b6c-4fb2-9002-7d41b4bc6b75

We can therefore remove this domain from this special text file
to keep it a bit smaller and understandable.