Add top level domains allowed to use Notify
There are certain top level domains that are used in the public
sector. Some/most of these are listed on
https://www.gov.uk/guidance/check-if-your-organisation-can-get-a-govuk-domain-name

I've reviewed the 14 domains listed and found that 11 of these are
either in this list already, or are attributed to an organisation
on Notify. This means that anyone signing up with an email address
that is a subdomain of those 11 will be able to use Notify
automatically.

This commit adds the 3 remaining top level domains. We've gotten
away without having these for quite a while so it won't be causing
us much pain (ie lots of requests to approve subdomains joining
Notify) but will hopefully help very slightly and makes this
list a bit more thorough.

From a security angle, we don't know the exact process that these
3 top level domains issue subdomains, but we are assuming it is
'good enough' (plus will be just one of the layers of defence to
stop malicious activity on Notify).