Adding validator for potentially malicious sender IDS
Following on from an incident, it was identified that we could
block senderids that were for companies, such as Evri etc.

This work is the initial effort in this regard, we will look at
expanding this list of companies and perhaps keeping it in the database
in later work.

The work is case insensitive for the sender IDs involved in the incident
as well as some banks.