Use hidden attribute in cookies page
Using this instead means we make explicit what's
hidden when the page loads, in the HTML. It also
means users won't see things they shouldn't if the
CSS doesn't load, which is often raised as a
requirement in accessibility accessments.

Having state in the HTML also helps you write
Python tests that check the HTML for an idea of
what's happening.