Make default action for all forms current URL
This should fix the issue we've seen whereby the
fragment identifier ('#' + some text) part of the
current URL is carried forward to the next page,
because `request.url` doesn't contain it.

Note: this doesn't do any further processing to
`request.url` so the action will be set to the
full URL, including protocol and domain, ie:

https://notify.works/<service_id>

Rather than:

/<service_id>

Flask should know the domain and the protocol
should carry over so I can't see this causing any
problems with our forms submitting correctly but
it is a bit ugly.