Re-enable the flake8 print check
We don’t want print statements in production code – should be using
logging instead.

Ruff doesn’t have this check on by default[1] so we need to explicitly
include it.

The existing `print` statements in this codebase seem to be there
intentionally (they are in a command not a public endpoint) so I’ve
marked them as exempt from checking.

***

1. https://docs.astral.sh/ruff/rules/print/