Change branding query parameter name
All of the branding routes use `branding_choice` as the name of the
query parameter apart from the branding options route, which was using
`branding_option` instead.

It makes sense to use `branding_choice` everywhere since it keeps things
consistent and is the name that the `_letter_branding_flow_query_params` and
`_email_branding_flow_query_params` helper functions use.

This change adds `branding_choice` as a query param while falling back
to looking for `branding_option` as a query param so nothing breaks for
people if they are in the middle of changing branding as this change is
deployed.