Add branding choice to backlinks after selecting a branding option
Once you have selected a letter or email branding option from the
initial page, the back link on the subsequent page should contain the
`branding_choice`. This is so that we can remember your option when
clicking "back". Some of the back links did contain `branding_choice`
but others didn't so this adds it to the back links that did not already
have it.