Remember branding choice when viewing branding form after clicking "back"
Now that the selected branding choice is always being passed from the
`letter_branding_options` and `email_branding_options` views to the next
page, and is also in the query parameter when returning to that page we
can use it to populate the form. When clicking "back" to return to the
branding options page, your previous choice is selected.