Remove unused branding query param
In order to rename the branding query parameter `branding_option` to
`branding_choice`, we first added `branding_choice` which had the same
value of `branding_option` and wrote the code to check both. This was
done so that the existing journey didn't break if someone was in the
middle of using it when the change was deployed.

Now that the change has been deployed for some days, `branding_option`
should not be being used so can be removed.