Split letter uploads tests so that each route is tested separately
This is so that in subsequent commits I can make changes to
just the letter upload page and not the letter attachment page.
My changes will break some of the assertions of what content
will be expected on the page. The equivalent changes may be made
to the letter attachment page later so these then might revert
back to being parametrized (maybe).