Shorten pdf read validation error message
The previous message was just long enough that we wanted to spread
it over 2 lines. We previously tried to have a <br> tag in it, but
this doesn't work with the error summary component. We did consider
putting in some non breaking spaces to have the inline validation
error spread over 2 lines and the error summary still on a single
line...

But in the end, we went for the easiest approach given this is
an uncommon error, which is to shorten it slightly. It now appears
as just a single line for both inline and error summary.