Move tel numbers validation & add summary messages
Moves the validation for phone numbers out to
validator classes, so they're the same as the
other validation logic used on pages with a
variable number of fields. I mean these:
- validation for domain names
- validation for email addresses

There are a few different places you can put this
logic in WTForms forms but this made sense for
this work, at least.

Other notes about these changes:

1. The error_summary_messages list

I've partially copied the existing WTForms
approach by having the error_summary_messages list
as an attribute on field objects, like the errors
list, and flushing it before validation. I stopped
short of also adding the errors put in
error_summary_messages to the parent FieldList or
Form instances due to not needing them.

2. The '%s' in error summary messages

In the messages written so far, we've needed to
change both the noun (ie. 'mobile number') and the
number of its field in the list of fields. Because
of this we leave this to the template. The
template then adds the noun (AKA 'thing') and
number. It assumes the noun is lowercase and will
capitalise the whole sentence.