Change content for template deletion warning hint
The query to check when template was last used would often time out
(Sentry has record of 10 000 errors this caused - and it's a
manual process to delete a template, so it's a lot of users seeing
error page :< ).

One measure to prevent that is for us to only search withing last
year - that should be good enough for the user to be warned
if they are deleting a well-used template by mistake.

So in this PR we are changing the content of the warning to match
the check we are actually performing in the API
(checking for latest template usage, if any, within last year).