bump utils to 77.2.1 (changes to redis delete)
this involved changes to the tests. Previously each call to delete would
emit one statement, now they emit two statements, both with the flag
raise_exception=True.

To fix the tests, added a couple of helper methods to a new Mock
subclass that we can use to assert "were these keys deleted, in any
order and any amount of times". There's two functions - one for checking
exactly what is deleted and the other for checking if the things you
care about were deleted but allow other things to be deleted too.