Show error-summary validation for attach pages upload
We only want to show the new error summary error messages for simple validation if
- form is submitted empty
- file is not a PDF
- PDF fie is bigger than 2MB

For any any fiel content validation, we want to keep old complex error messaging.

Attach pages route renders two different HTML templates: initial render (attach-pages.html) and any error messages and manage-attachment.html for any subsequent reuploads and error messages if initial upload was valid.

In order not to rewrite the journeys and introduce a new flow for complex content validation, we make use of an additional parameter to
a) set `error_summary_enabled` on `render_template` to `true` if there's a form file field error (the three scenarios described above) which in turn shows the error summary in the template
b) pass it to the template and conditionaly show/hide inline error message and new/old error-summary