Fix cancel button on scheduled jobs
When AJAX updates the page for a scheduled job the `<form>` is posting
to `/services/…/jobs/47a70916-c90f-477c-aac6-263a847321a1.json?status=`

It should be posting to `/services/…/jobs/…` (note no `.json`)

The `.json` endpoint doesn’t accept POST requests therefore is returning
`405`.

This was introduced by https://github.com/alphagov/notifications-admin/pull/5068/files
where we set the `action` of every form to the URL of the current
request.

This commit overrides this and forces the `action` to the correct URL.