Add cookie cleanup code
Adds a simple script that checks if GA and/or
`cookies_policy`` cookies are present.

If any are, remove them from user's browser as
they're no longer needed and it's good practice
to do so.

Adds test for said fucntion and updates test
cookie helpers to add `getCookie` helper and
remove `deleteCookie` helper.