Add aria-describedby part of error pattern
This adds the missing part of the Design System
error pattern, whereby we use aria-describedby to
link the error message to the text input.

This change doesn't allow for the text inputs to
already have aria-describedby set, to simplify the
JS code. This works for all our current usecases.

This commit also adds a test for this linking,
adds values to the text inputs in the test for
validation, so we're not just testing text inputs
with empty values but an error state.

Finally, this also stops a check on attributes
shared by all text inputs checking
aria-describedby, as it can't exist in the HTML
that loads initially.