Add internal note to Zendesk tickets from non-logged in users
If a ticket was created by someone who was not logged in to Notify, we
immediately add an internal note to it. This is because it is possible
to put down any email address in the support form if not logged in - we
want to know if the ticket was created by an authenticated user or not.