Fix letter rates
The letter rate for a letter that is first class and 5 sheets is 1.15p,
but was appearing on the pricing page as 1.14. This is due to the
inaccuracy of floating point numbers then being converted to an int,
which truncated the value. We can avoid this happening by rounding the
number first.