Remove reference to issue with validation
This referred to the following problems, picked up
in previous accessibility audits:
- validation UI not being consistent across pages
- validation UI (including semantics) not
  complying with the requirements of the 'recover
  from validation errors' pattern in the design
  system

See:
https://design-system.service.gov.uk/patterns/validation/
for last point.

Where these issues were identified, the auditors
recommended we just use the design system
validation components and pattern.

We've since updating all our pages that use
validation and can be covered by that pattern to
use the design system components. This means they
are all visually and semantically consistent.

We've also made all the changes listed in the
pattern not already covered by using the
components.

Because of all that, I'm removing this issue.