Add extra back-link option to email branding upload logo page
There are three ways you can reach the `email_branding_upload_logo`
view, and the back link logic for the view only took two of those
into account.

We had back links for the cases of being redirected there from the pages
to choose a banner type or a banner colour. However, if you have selected
the option of the GOV.UK logo and your own logo, you get taken to the page
from the choose logo page (the page to select either a government identity
logo or to upload a logo), so the back link should go back there.