Remember banner type selected when going back to add banner page
We want the radio buttons to remember the banner type selected when
going back to the add banner email branding page. To do this, we need to
ensure that the `brand_type` query parameter doesn't get removed when
going back to the page. We then use this value to check the appropriate
radio button.