Remove Emergency Alerts service type indicators
The only time we badge the service name now is when a service is
suspended:
https://github.com/alphagov/notifications-admin/blob/d94aea6bca2020ba83e6bdb4a5b7c7304214ff20/app/templates/service_navigation.html#L4-L6

The other colours and styles of badges were used for differentiating
Emergency Alerts services where it was very important to know which
channel and operator an alert was being sent on. Emergency Alerts is no
longer part of Notify so we don’t need these in our codebase.