Make width of empty table message match number of columns
When we show a ‘this table is empty’ message we want it to span the full
width of the table.

Previously we were doing this by setting `colspan` to an arbitrary high
number (10), which is higher than the number of columns any of our
tables have.

At some point this seems to have stopped having the desired effect, in
Chrome at least.

This commit fixes the problem by setting `colspan` to exactly the number
of columns the table has.