Create separate Javascript bundles
GOV.UK Frontend requires to run in script type='module'
and our own Javascript cannot at the moment,
so we need to create two bundles; and ES module
bundle and our own app bundle as before.

This requires changes to the Gulp tasks and file
entry points as well as script inclusions in the
template.