Update CSS class to show/hide JS ehnanced elements
With update to govuk-frontend v5 their javascript
adds `govuk-frontend-supported` class to the
`body` element only in the browsers they support
(no IE), but they also add a `js-enabled` for
legacy reasons.

We now follow their browsers support so our own
Javascript does a check for
`govuk-frontend-supported`class and if present,
executes. This means we can no longer rely on
the `js-enabled` (which we cannot remove) to
show/hide elements.

For now, until we can move and redo all of our
Javascript to ES modules, we'll hide/show elements
based on the above class.