Add headings to templates form page
For browsers and users that don't execute our
Javascript we have added helpful headings to a
long page, which we can hide when Javascript
executes and adds a class to the body as the
page actions make sense in context.

We have made slight tweaks to the content so that
the region label and page heading can share the
same text. This is to avoid unnecessary duplication
and potenctila divergence in the future.