Purge Redis cache when unsubscribe report completed
We also cache a summary of unsubscribe reports that the person running
the service has batched up.

When a user marks one of these reports as completed this information is
included in the summary. So we need to purge a cache.

Should probably have a test which covers the `POST` route here too