Make unsubscribe report titles friendlier
At the moment the title of our reports follow the same format no matter:
- the range they cover
- how recent they are

This commit makes the titles aware of their context (in other words
what time it is now and what other reports exist).

This makes them better in two situations especially:
- there are multiple reports on the same day (where we now add the time
  to help differentiate)
- there are very recent reports (where we now say today or yesterday
  instead of the numeric date)

The alternative would be to always show the time and year for every
report, but this would feel noisy and overly-specific in most cases.