Generate asset URLs at request time, not build time
In a small handful of cases – all where we have radio buttons with
images – we were generating the asset hashes for these images at the
time of starting the app.

If the images were missing (for example because the gulp task had not
finished running) this would cause an exception.

This commit changes the radios-with-images compnent to defer generating
the hashes until the `<img>` tags are rendered in the HTML. This is the
same way that all other asset URLs are rendered.

This means that the app can still start up, but an exception will still
be thrown at runtime or while running the (now more robust) tests if the
image is missing.