Make ‘mark as completed’ checkbox available as soon as unsubscribe report is downloaded
At the moment you can’t mark a report as completed until you have
downloaded it.

However when you click the ‘Download this report’ link you stay on the
same page while the report downloads. This means you have downloaded the
report but still can’t click the checkbox.

This commit changes the flow when clicking the download link to:
- batch up the requests into a report
- server-side redirect to the page for that report (checkbox now
  enabled)
- client-side redirect to the download (while the user stays on the same
  page)

So the effect – without any AJAX – is that the report is downloaded and
the checkbox is enabled simultaneously.