Use report `title` for confirmation message
Rather than hand-crafting the to/from dates here we can use the `title`
property introduced in https://github.com/alphagov/notifications-admin/pull/5160/files
to give a more context-dependent title for the report. It will also
match the title shown later down the page.

This also rewords the message slightly to be less verbose and passive.