Enable DS validation on tour pages
Enable error-summary Design System validation pattern on tour pages to match the rest of the admin.

As field's auto focus iterferes with the required error-summary autofocus logic was added to only autofucs when there's to form error.

Additionally, this also fixes the bug on one-off-send journey where autofocus should be set, but it isn't working