Improve display of unsubscribe report status
This makes the status align to the bottom of the row, same way we do on
the received text messages page (this means that the baseline of the
small text lines up which is neater).

It also makes the status grey if the report is not downloaded, which
feels right as this is an indeterminate state, like ‘sending’ is when
sending a message.