Use older_than to paginate notifications CSV report that services download
Paginating by page numbers made us use offset, which makes
large queries quite expensive: https://use-the-index-luke.com/no-offset

This caused issues for notifications CSV reports that services can
download. If a service has sent many notifications, the report will
be very big - even above 1,000,000 rows.

To prevent it, for this use case we will use older_than parameter
to paginate instead - that is - we get our notifications sorted
by created_at date. So if we set older_than parameter to the id
of the oldest notification in a current page/batch, we know
where to start our next batch without using the inefficient
offset mechanism.

We still track pages numbers for now - this helps with quick
implementation, but we should look at removing them as a
refactor in the next PR.