Setup Jest to build & test ES modules files
As new/updated Javascript functionality is written
as ES modules, Jest needs to be able to transform
them to CommonJS as we're not in a position to be
fully ESM yet.

To transpile we use babel-jest which needs uptodate
babel-core and babel-present-env.

In package.json Jest config we tell it to transpile
.mjs test files as any new ESM modules code does not
need the window object legacy Javascript is attached
to, so any new tests are written with .mjs suffix.

To configure Babel, babel.config.js is set-up and
targets 'test' env as that's what Jest sets.

Updated babel-core and preset-env have not
introduced any extra code into the built static
`all.js` and `all-esm.mjs files` (see PR screenshot)

To be able to transpile ESM files,
`node --experimental-vm-modules node_modules/.bin/jest`
needs to be added to the the test command. More at
https://jestjs.io/docs/ecmascript-modules