models: use _get_items method instead of unusual client_method pattern
having a class attribute that's a method bound to another class's
instance is fairly weird, and means the class is undefinable/
unimportable without the creation of the instance in question.

this starts to cause trouble when the construction of the api
clients starts to depend on the existence of an app context.
for the same reason this also switches to referencing
NON_CANCELLED_JOB_STATUSES on JobApiClient instead of a
magic global instance.