Enable error-summary valdation pattern on tour
Introduce the pattern when form submission
errors occur, to be in lien with the rest of the
website.

As field's auto focus interferes with the
required error-summary autofocus logic was added
to only autofucs when there's to form error.

Additionally, this also fixes the bug on
one-off-send journey where autofocus should be
set, but it isn't working.