For notifications CSV report, get older_than arg from last notification
in the batch.

Instead of getting older_than from API, generate_notifications_csv()
can find it itself - if notifications fill whole page, that means
there could be more to pull - so take id of last one in the batch
and pass it on when calling API next.

This is how our public API users do it, too, so it makes our
system more consistent.