Ensure Zendesk tickets about phishing senders are not sent to users
Setting the user details when creating the creating a Zendesk ticket
about potential phishing senders means that the ticket appears to come
from them and they also see the ticket. The ticket should be internal
only, so this removes the user details from the ticket object. The user
and user's service are still visible in the ticket note.