Remember "sender_id" on set-sender page if going "back"
The "set-sender" page was setting `session["sender_id"]` to `None`, but
if you are returning to the page having clicked "back" when sending, we want
to be able to remember your choice and to pre-populate the form with it.

This change now means that we don't clear the "sender_id" from the
session if you're going back to the set-sender page and instead
highlight the option that was previously selected.