TemplatePreview: convert into TemplatePreviewClient, follow other api client patterns
having this follow a completely different pattern to all our
other api clients (while also not appearing to admit that it's
an api client through its naming) is weird and causes problems
when trying to convince it to use a persistent requests.Session.

firstly this is no longer just a loose collection of
mostly-independent functions, but an actual class that expects
to be instantiated. this instance is used to hold the persistent
requests.Session.

the instance however needs to be treated like the other api
clients' instances and use LazyLocalGetter to construct/hold
thread-local copies.

like other api clients, it also no longer makes ad-hoc references
to current_app and current_service. config parameters that were
read from current_app are now set at construction time through
passing an explicit app and methods that need to refer to
current_service now take a service argument where that can be
supplied. this latter change was needed to avoid a painful
import loop problem with current_service.