Add announcement of 1 October price increas
We expect Royal Mail to increase their prices on 1 October.

We don’t know any details yet, but we need to announce this to our users as early as possible.

This PR adds an announcement to the top of the page. We need to deploy this before we send an email about the price change to our users.