API clients: refactor construction, use thread-local instances
using globals is not thread safe and will start to cause
problems in our api clients as soon as they start doing
things that involve persistent state. instead, we need to
use thread-local instances of these api clients, allowing
the internal design of those clients to be unconcerned
with thread contexts.

(side note, the NotifyAdminAPIClient-based clients already
use persistent state, they probably don't run into
[noticeable] problems at the moment [despite the admin app
using eventlet] because of the comparatively low
request-rate for admin and the fact that the requests are
effectively "sharded" across ~30 separate clients, each
with their own requests.Session, making collisions less
likely)

the flask-style `init_app()` pattern is particularly
awkward for keeping thread-local copies, as we need to be
able to construct new thread-local instances on-demand
(because we don't necessarily know how many threads we have
initially and threads can be re-created at any time). the
`init_app()` pattern however rather assumes all
initialization will happen at app-init time. it's also
awkward because it means a client instance has an "existing
but not yet configured/valid" state that makes it hard to
reason about or add sensible type-hints for.

so this moves all initialization to the constructor. that
however means we can't access these instances outside an
app context because they need to refer to current_app for
configuration. strongly binding them to one specific app
instance out-of-app-context would also have been a
possibility but it causes significant problems when things
like test fixtures start playing with multiple app contexts
(intentionally or unintentionally).

some amount of test refactoring was needed here to ensure
that tests had an app context when they were going to
access these api clients. a lot of this involved moving
any `mocker` fixtures to the last argument of the test
function, making them one of the first to be tidied up,
and hopefully before the app context teardown cleans up
old LazyLocalGetters.

in doing this, it became clear that some fixture mechanisms
(e.g. the test_navigation ones) had been misbehaving and
starting an app context during the import phase which never
got closed. meaning some of the tests had been implicitly
depending on being in an app context and therefore needed
fixing as soon as the state-leaking mechanisms were fixed.

this unfortunately is all bundled up with the
notifications-utils bump, which, along with introducing
the required LazyLocalGetter, includes breaking changes
to AntivirusClient and ZendeskClient, updating them to use
the new construction pattern and adding an internal
requests.Session, meaning we have to convert those clients
too in the same commit (or having an in-between commit
whose tests don't pass)