Delete cached statistics when processing report
Marking a report as completed now removes its requests from the count
returned to the dashboard: https://github.com/alphagov/notifications-api/pull/4184/files

However the actual number does not update because the previous value is
being cached in Redis.

This commit deletes the Redis cache any time we mark a report as
completed/not completed.