Send `user_created_at` field to Zendesk for authenticated users
We want to populate a new field on the Notify Zendesk form showing when
a Notify user account was created.

For users who are logged in when using the feedback form, we now pass
the UTC datetime that their Notify user account was created as an
argument to the support ticket. For non-logged in users we can just pass
`None` as the argument, which ensures the new date picker won't get
populated.

The `user_created_at` argument is optional.