Add type coercion for primitive types
Pretty straightforward, gives a guarantee that a property will be of a
certain type, or `None` if not present in the dictionary.

This is different from dataclasses which always require every property
to be present on the instance.

This preserves the existing behaviour, which is that if the underlying
JSON doesn’t contain the field then accessing the property raises an
`AttributeError`.