Change how we infer if there is a previous/next notifications page
So far we were doing this by checking for previous/next links
in the response from API, now instead we infer there is a previous
page if current one is not number 1, and there is most likely
a next page if current page is full (50 notifications).

Our new heurestics for assessing if there is a next page will
sometimes return a false positive (when there is exactly 50 notifications left,
so current batch was full, but next batch will come empty). But
we are willing to accept that, and to occasionally serve users
an empty last page, because the new way is much more resource-efficient.